# specs/005-safety/contracts/safety.openapi.yaml
openapi: 3.1.0
info:
  title: 오구오구 API - 위험 감지와 안전장치 (005 추가분)
  version: 0.5.0
  description: |
    005-safety에서 추가하고 바꾸는 경로와 스키마다. 구현 첫 작업에서 저장소 루트의
    contracts/openapi.yaml(누적 계약)에 합친다(003 research R13). 이 파일은 리뷰와 검증용으로
    단독으로도 유효하게 두며, ErrorEnvelope, Author, Cursor, Size, PostId, CommentId는 루트 계약의 정의와 같다.

    모든 경로는 인증이 필요하고, 온보딩 전 회원에게는 403 ONBOARDING_REQUIRED를 돌려준다.

    운영자 경로(/api/v1/operator/**)는 member.role이 OPERATOR인 회원만 쓴다. 운영자가 아니면
    경로가 있다는 것도 드러나지 않도록 404 NOT_FOUND를 돌려준다(research R10). 웹의 BFF 프록시는
    operator/**를 넘기지 않는다. 운영자는 API를 직접 부른다.

    기존 스키마의 변경(루트 계약에 합칠 때 반영한다):
    - PostDetail: 작성자에게만 safety(ContentSafety)가 실린다. 다른 회원에게는 필드가 없다.
    - Comment: hidden(boolean)을 더한다. 숨긴 댓글을 다른 회원이 보면 hidden=true, content와 author는 null이다.
      작성자 자신에게는 content와 author가 그대로이고 safety가 실린다.
    - FeedItem: 내가 쓴 글 목록에서만 hidden(boolean)이 true일 수 있다.
    - NotificationType: SUPPORT_NOTICE, CONTENT_RESTORED, REVIEW_KEPT를 더한다.
    - 글과 댓글의 본문, 미리보기는 보는 사람이 작성자가 아니면 욕설이 가려진 채로 나간다(research R6).
servers:
  - url: http://localhost:8080
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TargetType:
      name: targetType
      in: path
      required: true
      schema: { $ref: '#/components/schemas/TargetType' }
    TargetId:
      name: targetId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    ErrorEnvelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: false }
        data: { type: 'null' }
        error: { $ref: '#/components/schemas/ErrorResponse' }
    TargetType:
      type: string
      enum: [POST, COMMENT]
...
    SupportResource:
      type: object
      required: [name, phone, hours, description]
      properties:
        name: { type: string, examples: [자살예방상담전화] }
        phone: { type: string, examples: ['109'] }
        hours: { type: string, examples: [24시간] }
        description: { type: string }
    ReportReason:
      type: string
      description: DANGEROUS 위험해 보여요, ABUSIVE 욕설이나 비방, SPAM 광고나 도배, OTHER 기타
      enum: [DANGEROUS, ABUSIVE, SPAM, OTHER]
    ReportRequest:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType: { $ref: '#/components/schemas/TargetType' }
        targetId: { type: integer, format: int64 }
        reason: { $ref: '#/components/schemas/ReportReason' }
        detail:
          type: [string, 'null']
          description: reason이 OTHER일 때만 받는다. 200글자(사람이 보는 글자 단위)까지
    ReviewRequestBody:
      type: object
      required: [targetType, targetId]
      properties:
        targetType: { $ref: '#/components/schemas/TargetType' }
        targetId: { type: integer, format: int64 }
...
    OperatorNote:
      type: object
      properties:
        note:
          type: [string, 'null']
          description: 처리 기록에 남길 메모. 200자까지. 본문을 옮겨 적지 않는다
    ReportDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          description: RESOLVE는 처리됨으로, REJECT는 기각으로 닫는다. 대상을 숨기려면 hideContent를 따로 부른다
          enum: [RESOLVE, REJECT]
        note: { type: [string, 'null'] }
    ReviewDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          description: KEEP은 숨김 유지, RESTORE는 숨김 해제
          enum: [KEEP, RESTORE]
        note: { type: [string, 'null'] }
    TermKind:
      type: string
      description: CRISIS 위기 표현, CONCERN 우려 표현, PROFANITY 욕설, ALLOW 욕설이 들어 있어도 가리지 않을 낱말
      enum: [CRISIS, CONCERN, PROFANITY, ALLOW]
    Term:
      type: object
      required: [termId, kind, term]
      properties:
        termId: { type: integer, format: int64 }
        kind: { $ref: '#/components/schemas/TermKind' }
        term: { type: string, description: 정규화한 형태 }
    TermRequest:
      type: object
      required: [kind, term]
      properties:
        kind: { $ref: '#/components/schemas/TermKind' }
        term: { type: string, description: 2~20글자. 저장할 때 정규화한다 }
  responses:
    BadRequest:
      description: 입력 검증 실패 (INVALID_REQUEST)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: 인증 없음 또는 세션 만료, 무효 (UNAUTHORIZED, SESSION_EXPIRED)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: 온보딩 전 (ONBOARDING_REQUIRED)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: 없거나, 지웠거나, 숨긴 대상 (POST_NOT_FOUND, COMMENT_NOT_FOUND)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    OperatorOnly:
      description: 운영자가 아니거나 대상이 없음 (NOT_FOUND). 운영자가 아닌 회원에게 경로의 존재를 드러내지 않는다
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
  /api/v1/safety/support-resources:
    get:
      tags: [safety]
      operationId: getSupportResources
      summary: 도움 리소스 목록 (US1-AC1, AC3, AC4). 보이는 순서대로
      responses:
        '200':
          description: 도움 리소스
          content:
            application/json:
              schema:
                type: object
                required: [success, data, error]
                properties:
                  success: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/SupportResource' }
                  error: { type: 'null' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/reports:
    post:
      tags: [safety]
      operationId: reportContent
      summary: 글이나 댓글 신고 (US3-AC1~AC5)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportRequest' }
      responses:
        '204':
          description: 접수됨. 신고 수나 처리 상태는 돌려주지 않는다
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 온보딩 전(ONBOARDING_REQUIRED) 또는 자기 글이나 댓글(CANNOT_REPORT_OWN_CONTENT)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 이미 신고한 대상 (ALREADY_REPORTED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          description: 한 시간 20건 초과 (REPORT_RATE_LIMITED). Retry-After 헤더에 남은 초
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/v1/review-requests:
    post:
      tags: [safety]
      operationId: requestReview
      summary: 숨겨진 내 글이나 댓글의 재검토 요청 (US4-AC8). 대상마다 한 번
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewRequestBody' }
      responses:
        '204':
          description: 접수됨
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: 내 것이 아니거나, 숨겨지지 않았거나, 없는 대상 (POST_NOT_FOUND, COMMENT_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: 이미 요청함 (REVIEW_ALREADY_REQUESTED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /api/v1/operator/assessments/{assessmentId}/reviewed:
    parameters:
      - name: assessmentId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      tags: [operator]
      operationId: markAssessmentReviewed
      summary: 판정을 확인함으로 표시 (US4-AC1). 이미 확인했어도 204
      responses:
        '204':
          description: 표시됨
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
...
  /api/v1/operator/reports/{reportId}/decision:
    parameters:
      - name: reportId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      tags: [operator]
      operationId: decideReport
      summary: 신고 닫기 (US4-AC5). 이미 닫힌 신고면 그대로 204
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportDecision' }
      responses:
        '204':
          description: 닫힘
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
...
  /api/v1/operator/review-requests/{reviewId}/decision:
    parameters:
      - name: reviewId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      tags: [operator]
      operationId: decideReview
      summary: 재검토 닫기 (US4-AC3, AC9). RESTORE면 숨김을 풀고, KEEP이면 유지한다. 작성자에게 알림이 간다
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecision' }
      responses:
        '204':
          description: 닫힘
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
  /api/v1/operator/contents/{targetType}/{targetId}/hidden:
    parameters:
      - $ref: '#/components/parameters/TargetType'
      - $ref: '#/components/parameters/TargetId'
    put:
      tags: [operator]
      operationId: hideContent
      summary: 대상 숨기기 (US4-AC4). 그 대상의 열린 신고를 모두 처리됨으로 닫는다. 이미 숨겨져 있어도 204
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorNote' }
      responses:
        '204':
          description: 숨김
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
    delete:
      tags: [operator]
      operationId: unhideContent
      summary: 숨김 풀기 (US4-AC3). 작성자에게 알림이 간다. 열린 재검토 요청은 RESTORED로 닫는다. 숨겨져 있지 않아도 204
      responses:
        '204':
          description: 다시 보임
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
  /api/v1/operator/terms:
    get:
      tags: [operator]
      operationId: listTerms
      summary: 낱말 목록 (FR-002, FR-014)
      parameters:
        - name: kind
          in: query
          required: false
          schema: { $ref: '#/components/schemas/TermKind' }
      responses:
        '200':
          description: 낱말 목록
          content:
            application/json:
              schema:
                type: object
                required: [success, data, error]
                properties:
                  success: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Term' }
                  error: { type: 'null' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
    post:
      tags: [operator]
      operationId: addTerm
      summary: 낱말 더하기. 모든 인스턴스에 퍼지기까지 최대 30초. 이미 있으면 그대로 200
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TermRequest' }
      responses:
        '200':
          description: 더해진 낱말
          content:
            application/json:
              schema:
                type: object
                required: [success, data, error]
                properties:
                  success: { type: boolean, const: true }
                  data: { $ref: '#/components/schemas/Term' }
                  error: { type: 'null' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }
  /api/v1/operator/terms/{termId}:
    parameters:
      - name: termId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [operator]
      operationId: removeTerm
      summary: 낱말 빼기. 없어도 204
      responses:
        '204':
          description: 빠짐
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OperatorOnly' }